game-service/game_service/openai_utils.py
import os

from openai.lib.azure import AzureOpenAI


def get_openai():
    return AzureOpenAI(
        api_key=os.getenv("AZURE_OPENAI_API_KEY"),
        api_version="2024-02-01",
        azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT")
    )
